requirements.txt
Django==3.2.13
djangorestframework==3.14.0
django-allauth==0.52.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework-simplejwt==5.2.2
dj-rest-auth==4.0.0